tests/requirements_test.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
certifi==2024.8.30
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.5
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.7
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.8.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1

alembic==1.13.2
psycopg2==2.9.9
setuptools==74.1.2

pytest==8.3.2
pytest-asyncio==0.24.0
factory_boy==3.3.1
Faker==28.4.1

black==24.8.0
flake8==7.1.1
flake8-bugbear==24.8.19
isort==5.13.2
mypy==1.11.2
types-aiofiles==24.1.0.20240626
